Added Sweeper Detection and unified binary image generation
The recorder is now passing the binary image to the ball tracker and generating the binary difference image is now centralized as a function in recorder. \n Binary difference image generation is now also applied to the pins video to find the time when the sweeper comes down. This is measured as a continuous row of pixels in white (pixel has changed to the reference). \n Slight updates to the settings was neccessary and the live debugging image now also shows the debugging image from the pin video.